+++ api/admin/internal/logic/member/level/memberlevelupdatelogic.go
@@ -32,7 +32,7 @@ func (l *MemberLevelUpdateLogic) MemberLevelUpdate(req types.UpdateMemberLevelRe
 		Name:                  req.Name,
 		GrowthPoint:           req.GrowthPoint,
 		DefaultStatus:         req.DefaultStatus,
-		FreeFreightPoint:      int64(req.FreeFreightPoint),
+		FreeFreightPoint:      float32(req.FreeFreightPoint),
 		CommentGrowthPoint:    req.CommentGrowthPoint,
 		PriviledgeFreeFreight: req.PriviledgeFreeFreight,
 		PriviledgeSignIn:      req.PriviledgeSignIn,
