+++ api/admin/internal/logic/member/statistics/memberstatisticsinfoaddlogic.go
@@ -29,7 +29,7 @@ func NewMemberStatisticsInfoAddLogic(ctx context.Context, svcCtx *svc.ServiceCon
 func (l *MemberStatisticsInfoAddLogic) MemberStatisticsInfoAdd(req types.AddMemberStatisticsInfoReq) (*types.AddMemberStatisticsInfoResp, error) {
 	_, err := l.svcCtx.MemberStatisticsInfoService.MemberStatisticsInfoAdd(l.ctx, &umsclient.MemberStatisticsInfoAddReq{
 		MemberId:            req.MemberId,
-		ConsumeAmount:       int64(req.ConsumeAmount),
+		ConsumeAmount:       req.ConsumeAmount,
 		OrderCount:          req.OrderCount,
 		CouponCount:         req.CouponCount,
 		CommentCount:        req.CommentCount,
