+++ api/front/internal/logic/order/canceluserorderlogic.go
@@ -80,8 +80,8 @@ func (l *CancelUserOrderLogic) CancelUserOrder(req *types.CancelUserOrderReq) (*
 
 	//5.返还使用积分
 	member, _ := l.svcCtx.MemberService.QueryMemberById(l.ctx, &umsclient.MemberByIdReq{Id: memberId})
-	i := member.Integration + integration
-	_, err = l.svcCtx.MemberService.UpdateMemberIntegration(l.ctx, &umsclient.UpdateMemberIntegrationReq{Id: memberId, Integration: i})
+	i := member.Integration + int32(integration)
+	_, err = l.svcCtx.MemberService.UpdateMemberIntegration(l.ctx, &umsclient.UpdateMemberIntegrationReq{Id: memberId, Integration: int64(i)})
 	if err != nil {
 		return nil, err
 	}
